+++ crates/teloxide-core/src/types/transaction_partner.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::{Gift, PaidMedia, RevenueWithdrawalState, Seconds, User};
+use crate::types::{Chat, Gift, PaidMedia, RevenueWithdrawalState, Seconds, User};
 
 /// This object describes the source of a transaction, or its recipient for
 /// outgoing transactions.
@@ -12,6 +12,7 @@ use crate::types::{Gift, PaidMedia, RevenueWithdrawalState, Seconds, User};
 pub enum TransactionPartner {
     Fragment(TransactionPartnerFragment),
     User(Box<TransactionPartnerUser>),
+    AffiliateProgram(TransactionPartnerAffiliateProgram),
     TelegramAds,
     TelegramApi(TransactionPartnerTelegramApi),
     Other,
@@ -34,6 +35,10 @@ pub struct TransactionPartnerUser {
     /// Information about the user.
     pub user: User,
 
+    /// Information about the affiliate that received a commission via this
+    /// transaction
+    pub affiliate: Option<AffiliateInfo>,
+
     /// Bot-specified invoice payload.
     pub invoice_payload: Option<String>,
 
@@ -48,6 +53,47 @@ pub struct TransactionPartnerUser {
 
     /// The gift sent to the user by the bot
     pub gift: Option<Gift>,
+}
+
+/// Contains information about the affiliate that received a commission via this
+/// transaction.
+#[derive(Clone, Debug)]
+#[derive(PartialEq, Eq, Hash)]
+#[derive(Serialize, Deserialize)]
+pub struct AffiliateInfo {
+    /// The bot or the user that received an affiliate commission if it was
+    /// received by a bot or a user
+    affiliate_user: Option<User>,
+
+    /// The chat that received an affiliate commission if it was received by a
+    /// chat
+    affiliate_chat: Option<Chat>,
+
+    /// The number of Telegram Stars received by the affiliate for each 1000
+    /// Telegram Stars received by the bot from referred users
+    commission_per_mille: u32,
+
+    /// Integer amount of Telegram Stars received by the affiliate from the
+    /// transaction, rounded to 0
+    amount: i32, // Can be negative for refunds
+
+    /// The number of 1/1000000000 shares of Telegram Stars received by the
+    /// affiliate
+    nanostar_amount: Option<i32>, // Can be negative for refunds
+}
+
+/// Describes the affiliate program that issued the affiliate commission
+/// received via this transaction.
+#[derive(Clone, Debug)]
+#[derive(PartialEq, Eq, Hash)]
+#[derive(Serialize, Deserialize)]
+pub struct TransactionPartnerAffiliateProgram {
+    /// Information about the bot that sponsored the affiliate program
+    pub sponsor_user: Option<User>,
+
+    /// The number of Telegram Stars received by the bot for each 1000 Telegram
+    /// Stars received by the affiliate program sponsor from referred users
+    pub commission_per_mille: u32,
 }
 
 /// Describes a transaction with payment for paid broadcasting.